+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 ### What's Changed
 * Added: JKBMS CAN - Extended protocol with version V2 by @Hooorny and @mr-manuel
+* Added: Min/Max lifetime temperature to history class and battery template by @mr-manuel
 * Added: Signal handler for clean service restart/shutdown by @mr-manuel
 * Changed: A lot of under the hood optimizations by @mr-manuel
 * Changed: Daly CAN - Refactored driver to match new CAN standard by @mr-manuel